+++ Products/CMFPlone/traversal.py
@@ -1,6 +1,7 @@
 from plone.resource.traversal import ResourceTraverser
 from zope.component import getUtility
 from plone.resource.interfaces import IResourceDirectory
+from plone.resource.interfaces import IUniqueResourceRequest
 from Products.CMFPlone.interfaces.resources import (
     OVERRIDE_RESOURCE_DIRECTORY_NAME)
 from zope.globalrequest import getRequest
@@ -19,12 +20,23 @@ class PloneBundlesTraverser(ResourceTraverser):
         resource_path = req.environ['PATH_INFO'].split('++plone++')[-1]
         resource_name, resource_filepath = resource_path.split('/', 1)
 
+        # If we're traversing through the ++unique++ namespace to indicate a
+        # stable resource, use the un-uniqued file path
+        is_stable_resource = resource_filepath.startswith('++unique++')
+        is_uniqued = IUniqueResourceRequest.providedBy(req)
+        if is_stable_resource:
+            resource_filepath = '/'.join(resource_filepath.split('/')[1:])
+
         persistentDirectory = getUtility(IResourceDirectory, name="persistent")
         directory = None
         if OVERRIDE_RESOURCE_DIRECTORY_NAME in persistentDirectory:
             container = persistentDirectory[OVERRIDE_RESOURCE_DIRECTORY_NAME]
             if resource_name in container:
                 directory = container[resource_name]
+                if (is_stable_resource and not is_uniqued and
+                        resource_filepath in directory):
+                    # Continue traversal if we have not uniqued the request
+                    return directory
                 try:
                     return directory[resource_filepath]
                 except: